+++ modules/DGM/InferTRW.cpp
@@ -29,7 +29,7 @@ namespace DirectGraphicalModels
 			for (size_t e_t : node->to) {
 				Edge *edge_to = getGraphPairwise().m_vEdges[e_t].get();
 				if (edge_to->node1 > edge_to->node2) continue;
-				float *msg = getMessage(e_t);
+				double *msg = getMessage(e_t);
 				for (byte s = 0; s < nStates; s++) node->Pot.at<float>(s, 0) *= msg[s];
 			}
 
@@ -44,8 +44,8 @@ namespace DirectGraphicalModels
 	void CInferTRW::calculateMessages(unsigned int nIt)
 	{
 		const    byte	  nStates	= getGraph().getNumStates();										// number of states
-		float			* data		= new float[nStates];
-		float			* temp		= new float[nStates];
+		double			* data		= new double[nStates];
+		double			* temp		= new double[nStates];
 
 		// main loop
 		for (unsigned int i = 0; i < nIt; i++) {										// iterations
@@ -55,13 +55,13 @@ namespace DirectGraphicalModels
 	#endif
 			// Forward pass
 			std::for_each(getGraphPairwise().m_vNodes.begin(), getGraphPairwise().m_vNodes.end(), [&](ptr_node_t &node) {
-				memcpy(data, node->Pot.data, nStates * sizeof(float));					// data = node.pot
+				memcpy(data, node->Pot.data, nStates * sizeof(double));					// data = node.pot
 
 				int	nForward = 0;
 				for (size_t e_t : node->to) {
 					Edge *edge_to = getGraphPairwise().m_vEdges[e_t].get();
 					if (edge_to->node1 > edge_to->node2) continue;
-					float *msg = getMessage(e_t);
+					double *msg = getMessage(e_t);
 					for (byte s = 0; s < nStates; s++) data[s] *= msg[s];				// data = node.pot * edge_to.msg
 					nForward++;
 				} // e_t
@@ -70,12 +70,12 @@ namespace DirectGraphicalModels
 				for (size_t e_f : node->from) {
 					Edge *edge_from = getGraphPairwise().m_vEdges[e_f].get();
 					if (edge_from->node1 > edge_from->node2) continue;
-					float *msg = getMessage(e_f);
+					double *msg = getMessage(e_f);
 					for (byte s = 0; s < nStates; s++) data[s] *= msg[s];				// data = node.pot * edge_to.msg * edge_from.msg
 					nBackward++;
 				} // e_f
 
-				for (byte s = 0; s < nStates; s++) data[s] = static_cast<float>(fastPow(data[s], 1.0f / MAX(nForward, nBackward)));
+				for (byte s = 0; s < nStates; s++) data[s] = static_cast<double>(fastPow(data[s], 1.0f / MAX(nForward, nBackward)));
 
 				// pass messages from i to nodes with higher m_ordering
 				for (size_t e_t : node->to) {
@@ -86,13 +86,13 @@ namespace DirectGraphicalModels
 
 			// Backward pass
 			std::for_each(getGraphPairwise().m_vNodes.rbegin(), getGraphPairwise().m_vNodes.rend(), [&](ptr_node_t &node) {
-				memcpy(data, node->Pot.data, nStates * sizeof(float));							// data = node.pot
+				memcpy(data, node->Pot.data, nStates * sizeof(double));							// data = node.pot
 
 				int	nForward = 0;
 				for (size_t e_t : node->to) {
 					Edge *edge_to = getGraphPairwise().m_vEdges[e_t].get();
 					if (edge_to->node1 > edge_to->node2) continue;
-					float *msg = getMessage(e_t);
+					double *msg = getMessage(e_t);
 					for (byte s = 0; s < nStates; s++) data[s] *= msg[s];
 					nForward++;
 				} // e_t
@@ -101,16 +101,16 @@ namespace DirectGraphicalModels
 				for (size_t e_f : node->from) {
 					Edge *edge_from = getGraphPairwise().m_vEdges[e_f].get();
 					if (edge_from->node1 > edge_from->node2) continue;
-					float *msg = getMessage(e_f);
+					double *msg = getMessage(e_f);
 					for (byte s = 0; s < nStates; s++) data[s] *= msg[s];
 					nBackward++;
 				} // e_f
 
 				// normalize data
-				float max = data[0];
+				double max = data[0];
 				for (byte s = 1; s < nStates; s++) if (max < data[s]) max = data[s];
 				for (byte s = 0; s < nStates; s++) data[s] /= max;
-				for (byte s = 0; s < nStates; s++) data[s] = static_cast<float>(fastPow(data[s], 1.0f / MAX(nForward, nBackward)));
+				for (byte s = 0; s < nStates; s++) data[s] = static_cast<double>(fastPow(data[s], 1.0f / MAX(nForward, nBackward)));
 
 				// pass messages from i to nodes with smaller m_ordering
 				for (size_t e_f : node->from) {
@@ -125,7 +125,7 @@ namespace DirectGraphicalModels
 	}
 
 	// Updates edge->msg = F(data, edge.Pot)
-	void CInferTRW::calculateMessage(float *msg, Edge &edge, float *temp, float *data)
+	void CInferTRW::calculateMessage(double *msg, Edge &edge, double *temp, double *data)
 	{
 		const byte nStates = getGraph().getNumStates();
 
@@ -133,16 +133,16 @@ namespace DirectGraphicalModels
 
 		for (byte y = 0; y < nStates; y++) {
 			float *pPot = edge.Pot.ptr<float>(y);
-			float max = temp[0] * pPot[0];																// vMin = tmp + edge.Pot(0, kdest)
+			double max = temp[0] * pPot[0];																// vMin = tmp + edge.Pot(0, kdest)
 			for (byte x = 1; x < nStates; x++) {
-				float val = temp[x] * pPot[x];
+				double val = temp[x] * pPot[x];
 				if (max < val) max = val;
 			}
 			msg[y] = max;
 		}
 
 		// Normalization
-		float max = msg[0];
+		double max = msg[0];
 		for (byte s = 1; s < nStates; s++) if (max < msg[s]) max = msg[s];
 		for (byte s = 0; s < nStates; s++) msg[s] /= max;
 	}
